% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileFunctions.R
\name{compileFunctions}
\alias{compileFunctions}
\title{compileFunctions}
\usage{
compileFunctions(object = parent.frame())
}
\arguments{
\item{object}{nested object like environment / list / vector}
}
\description{
wrapper of cmpfun function in compiler package to compile functions in a nested object like environment, list or vector
}
\keyword{internal}
