% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logMsg.R
\name{logMsg}
\alias{logMsg}
\title{logMsg}
\usage{
logMsg
}
\arguments{
\item{...}{Message objects}
}
\description{
ligthweigth function to write msg into a connection
}
\details{
default connection is stderr and the first argument for the title
}
\keyword{internal}
