% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalMsgId.R
\docType{data}
\name{.msgHandler}
\alias{.msgHandler}
\title{.msgHandler}
\format{
An object of class \code{character} of length 7.
}
\usage{
.msgHandler
}
\arguments{
\item{expr}{Expression}
}
\description{
Constant Vector for all message functions
}
\keyword{internal}
